+++ example/reflection_factory_example.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/2.1.2
+// BUILDER: reflection_factory/2.1.3
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -17,7 +17,7 @@ typedef __TI<T> = TypeInfo<T>;
 typedef __PR = ParameterReflection;
 
 mixin __ReflectionMixin {
-  static final Version _version = Version.parse('2.1.2');
+  static final Version _version = Version.parse('2.1.3');
 
   Version get reflectionFactoryVersion => _version;
 
@@ -52,7 +52,7 @@ class User$reflection extends ClassReflection<User> with __ReflectionMixin {
   }
 
   @override
-  Version get languageVersion => Version.parse('2.18.0');
+  Version get languageVersion => Version.parse('2.19.0');
 
   @override
   User$reflection withObject([User? obj]) =>